refactor(offline): move useIsOnline + getIsOnline to canonical homes
Per review feedback: these primitives are general browser helpers, not
offline-specific. Split the file:

- `src/lib/utils/browser/getIsOnline/getIsOnline.{ts,test.ts}` — the
  imperative `navigator.onLine` reader used outside React (e.g. by
  AvaQueryClient's `staleTime` / `retry`).
- `src/lib/hooks/browser/useIsOnline/useIsOnline.{ts,test.ts}` — the
  `useSyncExternalStore`-backed reactive hook used in components.

Test for `getIsOnline` is split out into its own file. Imports updated
in OfflineIndicator, AvaQueryClient, and useOfflineGate. The old
`src/lib/offline/useIsOnline.{ts,test.ts}` are deleted.
