+++ src/lib/utils/browser/getIsOnline/getIsOnline.ts
@@ -1,0 +1,4 @@
+/** Imperative online check for code outside React. */
+export function getIsOnline(): boolean {
+  return navigator.onLine;
+}
